fix(types): convert header expectation to duck-typed interface
some web-frameworks like opine come with not the standard request object,
but one that is wrapped with a framework pendent.

Such wrappers often add additional properties to the request, which results in
a type mismatch because the types are too different.

Relying on duck typing here makes the lib more compatible for typescript projects